% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.rdf.R
\name{as.rdf}
\alias{as.rdf}
\title{Change it as rdf connenction}
\usage{
as.rdf(con)
}
\arguments{
\item{con}{a tiny connection}
}
\value{
a con for rdf
}
\description{
Change the connection type into rdf database
}
\examples{
\dontrun{
con <- odbcConnect('tiny')
}
}
