% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R, R/backtest.plot.R
\docType{class}
\name{backtest-class}
\alias{backtest-class}
\alias{summaryStats}
\alias{naCounts}
\alias{totalCounts}
\alias{ci}
\alias{counts}
\alias{marginals}
\alias{means}
\alias{turnover}
\alias{summary,backtest-method}
\alias{summaryStats,backtest-method}
\alias{means,backtest-method}
\alias{counts,backtest-method}
\alias{totalCounts,backtest-method}
\alias{marginals,backtest-method}
\alias{naCounts,backtest-method}
\alias{turnover,backtest-method}
\alias{ci,backtest-method}
\alias{plot,backtest,missing-method}
\title{backtest class}
\usage{
\S4method{summary}{backtest}(object)

\S4method{summaryStats}{backtest}(object)

\S4method{means}{backtest}(object)

\S4method{counts}{backtest}(object)

\S4method{totalCounts}{backtest}(object, low.high.only = FALSE)

\S4method{marginals}{backtest}(object)

\S4method{naCounts}{backtest}(object)

\S4method{turnover}{backtest}(object, mean = FALSE)

\S4method{ci}{backtest}(object)

\S4method{plot}{backtest,missing}(x, type = "return")
}
\arguments{
\item{object}{A backtest object}

\item{low.high.only}{A logical value indicating whether to omit
all but the highest and lowest buckets.}

\item{mean}{A logical value indicating whether mean turnover should be appended}

\item{x}{A backtest object}

\item{type}{A character equal to "return", "turnover", "cumreturn", or "cumreturn.split"}
}
\description{
Contains results from the backtest function.
}
\details{
The primary method for accessing the \code{backtest} results is through
  the \code{summary} method.  \code{summary} provides different displays
  depending on the type of \code{backtest} object.  These displays are
  shown in the examples section.  Accessor methods such as \code{means},
  \code{counts}, \code{marginals}, \code{naCounts}, \code{turnover}, and
  \code{ci} may be used to extract other types of information from the object.

  A \code{backtest} object with a \code{natural} value of TRUE may be
  graphed by calling the \code{plot} method.  The default \code{plot}
  method graphs return.  The other plots, turnover and
  cumulative return, must be explicitly specified as \code{plot(object,
    type = "turnover")} or \code{plot(object, type = "cumreturn")}.

  The \code{backtest} object does not store the data frame used to create
  the \code{backtest.}  It only stores the results and the names of the
  vectors used in calculating these results.

  The results of a \code{backtest} are stored in a 5-dimensional array,
  \code{results}.  The 1st dimension contains one value for every element
  of \code{ret.var}.  The 2nd dimension contains one value for
  every element of \code{in.var}.  The 3rd dimension contains one value
  for every element in \code{1:buckets[1]}, a vector from 1 through the
  number of \code{by.var} buckets.  The 4th dimension contains one value
  for every element in \code{1:buckets[2]}, a vector from 1 through the
  number of \code{in.var} buckets.  The 5th dimension contains 4
  elements: \code{means}, \code{counts}, \code{trim.means}, and
  \code{NAs}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{summary}: Prints a header listing the .vars, then prints the table returned
by the "spreads" method.  Returns a list of result tables

\item \code{summaryStats}: Returns a data frame summarizing the results of the backtest.  The
entries of the data frame contain means in cases 1, 2, and 4, and
spreads in cases 3 and 5.  When a table of means is returned, the
".bt.spread" function is called and summary spread data is attached
to the right side of the data frame.  If a date.var is used,
".bt.mean" is called and mean summary data are attached to bottom
of the data frame.

\item \code{means}: Returns a list of matrices, with one matrix for each in.var,
containing the means data.

\item \code{counts}: Returns a list of matrices, one matrix for each in.var, containing
the counts data

\item \code{totalCounts}: This function computes the counts of non-NA values that went into
the calculation of spreads displayed by backtests's summary()
function. It is different from counts because it displays the sum
of counts from all buckets (or lowest and highest only), thus
allowing for output that matches the format of spreads output.

\item \code{marginals}: Same as the "counts" method, except that the total counts for each
row and column are added to the margins.

\item \code{naCounts}: Returns a list of matrices, one matrix for each in.var, containing
the NAs data

\item \code{turnover}: Returns the turnover for natural portfolios.  Passing a "mean"
argument will append the mean of the turnover(s) as the last row of
the matrix

\item \code{ci}: Returns the confidence intervals for each spread for all cases
except: (multiple in.var & any by.var) or (multiple in.var & multiple
ret.var).  Must vectorize or loop to get these.

\item \code{plot}: Plots returns, cumulative returns, or turnover, when passed a
\code{type} argument of \code{return}, \code{cumreturn}, or
\code{turnover}, respectively.
}}

\section{Slots}{

\describe{
\item{\code{in.var}}{Object of class \code{"character"} specifying
the \code{in.var} values for this backtest.}

\item{\code{ret.var}}{Object of class \code{"character"} containing
the \code{ret.var} values for this backtest.}

\item{\code{by.var}}{Object of class \code{"character"} containing
the \code{by.var}, if specified, for this backtest.}

\item{\code{date.var}}{Object of class \code{"character"} containing
the \code{date.var}, if specified, for this backtest.}

\item{\code{buckets}}{Object of class \code{"numeric"} containing
the number(s) of buckets used create quantiles from the
\code{in.var} and \code{by.var} values.}

\item{\code{natural}}{Object of class \code{"logical"} expressing
whether or not the intervals between observations, as specified by
\code{date.var}, and returns, as specified by \code{ret.var},
match.  If the interval between dates is one month, the interval
between returns should also be one month.}

\item{\code{do.spread}}{Object of class \code{"logical"}. If TRUE the \code{summary}
method displays information about the spread between the extreme
quantiles.  If FALSE this information is suppressed.  Defaults to TRUE.}

\item{\code{by.period}}{Object of class \code{"logical"}. If TRUE the quantiles are
recalculated within each date period.  If FALSE the quantiles are
calculated all at once.  Defaults to TRUE.}

\item{\code{overlaps}}{An object of class \code{"numeric"} which specifies
the number of prior periods to include in the current period's
portfolio weights calculation. If \code{overlaps} is the default of 1,
backtest behaves as usual and only uses a periods own data to
determine its portfolio.  If \code{overlaps} is set to
n > 1, a period's portfolio comprises the weighted mean of
portfolio weights from the previous n periods, with period n
having a weight of 1/n.}

\item{\code{id.var}}{Object of class \code{"character"} specifying
the \code{id.var} value for this backtest.}

\item{\code{universe}}{Object of class \code{"logical"} with length equal to the input data, with values TRUE if the corresponding row is included in the analysis and FALSE otherwise.}

\item{\code{data.original}}{Object of class \code{"data.frame"} containing the original input data.}

\item{\code{calculated.data}}{Object of class \code{"environment"} containing the delayed assignment promises for the "results", "turnover" and "ret.stats" variables.}
}}

\examples{
data(starmine)
bt <- backtest(starmine, in.var = "smi", ret.var = "ret.0.1.m", by.period = FALSE)

## Summary for a pooled backtest

summary(bt)

## A natural backtest

bt <- backtest(starmine,
  in.var = "smi", ret.var = "ret.0.1.m",
  date.var = "date", id.var = "id", natural = TRUE, by.period = FALSE
)

## Summary for a natural backtest

summary(bt)

## Other access methods

means(bt)
counts(bt)
marginals(bt)
naCounts(bt)

## Plotting methods
\dontrun{
plot(bt, type = "turnover")
plot(bt, type = "return")
plot(bt, type = "cumreturn")
plot(bt, type = "cumreturn.split")
}
}
\seealso{
\code{\link{backtest}}
}
