\name{backtest-package}
\alias{backtest-package}
\docType{package}
\title{
Exploring portfolio-based conjectures about financial
instruments
}
\description{
The backtest package provides facilities for exploring
portfolio-based conjectures about financial instruments
(stocks, bonds, swaps, options, et cetera).
}
\details{
\tabular{ll}{
Package: \tab backtest\cr
Type: \tab Package\cr
Version: \tab 0.4\cr
Date: \tab 2018-08-24\cr
Depends: \tab R (>= 2.3.0), methods, dplyr, gridExtra, tidyr, ggplot2\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}

Index:
\preformatted{
backtest                Creating an Object of Class Backtest
backtest-class          Class "backtest"
backtest.update         Modify a backtest's parameters
starmine                StarMine Rankings, 1995
}

Further information is available in the following vignettes:
\tabular{ll}{
\code{backtest} \tab Using the backtest package (source, pdf)\cr
}
}
\author{
Jeff Enos <jeff@kanecap.com> and David Kane <dave@kanecap.com>,
with contributions from Kyle Campbell
<kyle.w.campbell@williams.edu>, Daniel Gerlanc
<daniel@gerlanc.com>, Aaron Schwartz
<Aaron.J.Schwartz@williams.edu>, Daniel Suo
<danielsuo@gmail.com>, Alexei Colin <acolin@fas.harvard.edu>,
Luyi Zhao <luyizhao@gmail.com>, and David Liu <davjliu@gmail.com>

Maintainer: David Liu <davjliu@gmail.com>
}
\keyword{ package }
