% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/backtest.compute.R
\name{backtest.compute}
\alias{backtest.compute}
\title{Returns Backtest Object}
\usage{
backtest.compute(x, in.factor, ret.var, by.var = NULL, date.var = NULL,
  natural = FALSE, by.specified = NULL, do.spread = TRUE, id.var = NULL,
  by.period = TRUE, overlaps = 1)
}
\value{
a backtest object
}
\description{
\code{backtest.compute} returns a backtest object
}

