% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\docType{package}
\name{backtest}
\alias{backtest}
\alias{backtest-package}
\title{make a stock acount for back test}
\description{
make a stock acount for initial parameter, then buy or sell stock with function
}
