% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{backtest.update}
\alias{backtest.update}
\title{Update existing backtest objects with new parameters}
\usage{
backtest.update(bt, x, ...)
}
\arguments{
\item{bt}{A backtest object to act as the template}

\item{x}{An optional \code{data.frame} to replace the data used in \code{bt}}

\item{...}{Any number of arguments that are accepted by the backtest function.
Any arguments in this list override the parameters of \code{bt}}
}
\value{
Returns an object of class \code{backtest}
}
\description{
Takes an existing backtest object and applies new parameters or data.
}
\examples{
data(starmine)
bt <- backtest(starmine,
  in.var = smi,
  ret.var = c(ret.0.1.m, ret.0.6.m), by.period = FALSE
)
summary(bt)

bt2 <- backtest.update(bt, ret.var = ret.0.6.m)
summary(bt2)
## bt2 is identical to creating backtest(starmine, in.var = smi,
##   ret.var = ret.0.6.m, by.period = FALSE)

bt3 <- backtest.update(bt2, by.period = TRUE, date.var = date)
summary(bt3)
}
