% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc.turnover.R
\name{calc.turnover}
\alias{calc.turnover}
\title{Calculates the turnover}
\usage{
calc.turnover(x, portfolio.factor, date.factor)
}
\arguments{
\item{x}{is a character vector of unique identifiers}

\item{portfolio.factor}{is a factor created from a vector of numeric values as specified by an \code{in.var}}

\item{date.factor}{is a factor created from a vector of dates as specified by \code{date.var}}
}
\value{
the numeric turnover
}
\description{
\code{calc.turnover} Calculates the turnover for a long-short portfolio corresponding to a backtest.
}
\details{
This measure of turnover ignores price changes
}

