% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/categorize.R
\name{categorize}
\alias{categorize}
\title{Returns a factor corresponding to vector \code{x}}
\usage{
categorize(x, n, is.date = FALSE)
}
\arguments{
\item{x}{is a numeric vector}

\item{n}{is the denominator of the increase factor}

\item{is.date}{boolean value}
}
\value{
a factor corresponding to vector \code{x}
}
\description{
\code{categorize} returns a factor corresponding to vector \code{x}
}

