% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/backtest.R
\docType{methods}
\name{ci,backtest-method}
\alias{ci,backtest-method}
\title{The \code{ci} method returns a matrix of confidence intervals for spreads}
\usage{
\S4method{ci}{backtest}(object)
}
\description{
The \code{ci} method returns a matrix of confidence intervals for spreads
}

