% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ci.R
\docType{methods}
\name{ci,backtest-method}
\alias{ci,backtest-method}
\title{\code{ci} Method for Confidence Interval}
\usage{
\S4method{ci}{backtest}(object)
}
\arguments{
\item{object}{is the backtest object to work on}
}
\value{
a matrix of confidence intervals for spreads
}
\description{
The \code{ci} method returns a matrix of confidence intervals for spreads
}
\examples{
data(starmine)
bt <- backtest(starmine, in.var = "smi", ret.var = "ret.0.1.m", by.period = FALSE)
ci(bt)
}
\seealso{
\code{\link{backtest}} and \code{\link{backtest-class}} and \code{\link{summary}}
}

