% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_scale.R
\name{extreme_scale}
\alias{extreme_scale}
\title{extreme scale}
\usage{
extreme_scale(x, ex_num = 5)
}
\arguments{
\item{x}{vector}

\item{ex_num, }{int, default is 5}
}
\description{
rescale the extreme data
}
\examples{
extreme_scale(x)

}
