% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_acount.R
\docType{class}
\name{factor_acount}
\alias{factor_acount}
\title{factor_acount}
\format{An \code{\link{R6Class}} generator object
\preformatted{
 stock_acount$new(con, buy_stock, fee = 0.005)
 factor_update(buy_num, begin_date, end_date, wait_sell = F, refresh = T, show_trace = T)
}}
\usage{
factor_acount
}
\arguments{
\item{con}{a tiny connection}

\item{fee}{float, the fee you sold stock}

\item{begin_date, }{the first date to buy stock list}

\item{end_date, }{the end date to hold stock list}

\item{wait_sell, }{logical. if True, sell the stock sell every day}

\item{...}{the paramater inherit from stock_acount}
}
\description{
create a R6 object of acount to analyse factor. can calculate the factor/number rank stock easily.
\code{factor_update} update the history date
}
\details{
it inherit from stock_acount
}
\examples{
\dontrun{
con <- odbcConnect('tiny')
my_acount <- stock_acount$new(con = con, acount_f = 1000000)
my_acount$order_buy(as.Date('2015-01-04'), stock = 'SH600000',num = 1000)
my_acount$acount_update(as.Date('2015-01-04'))
my_acount$order_sell(as.Date('2015-01-05'), stock = 'SH600000', num = 1000)
my_acount$acount_update(as.Date('2015-01-05'))
my_acount$show_acount_f()
}

}
\keyword{data}
