% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deposit_rate.R
\name{get_deposit_rate}
\alias{get_deposit_rate}
\title{Get future deposit rate}
\usage{
get_deposit_rate(con, wind_code)
}
\arguments{
\item{con}{a data connection}

\item{wind_code}{vecter of future code list in wind_code format}
}
\value{
data.frame with wind_code, deposit_rate
}
\description{
Get future deposit rate
}
\examples{
\dontrun{
get_price(con, c('600000.SH','000001.SH'), 20110104, type = 'close')
}

}
