% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_price.R, R/get_price_pd.R
\name{get_price}
\alias{get_price}
\alias{get_price.default}
\alias{get_price.tiny}
\alias{get_price.rdf}
\alias{get_price_pd.default}
\alias{get_price_pd.rdf}
\title{Get stock price for one day}
\usage{
get_price(...)

\method{get_price}{default}(...)

\method{get_price}{tiny}(con, stocks, buz_day, ...)

\method{get_price}{rdf}(con, stocks, buz_day, type = "close")

\method{get_price_pd}{default}(...)

\method{get_price_pd}{rdf}(con, stocks, begin_date, end_date)
}
\arguments{
\item{con}{a tiny connection}

\item{stocks}{vecter of stock code list in tiny format}

\item{buz_day}{integer or character format like \%Y\%m\%d, test for this day}

\item{type}{choose price type as close, vwap}
}
\value{
a data.frame with stock code and price, trade_dt
}
\description{
Get stock price for one day
}
\examples{
\dontrun{
con <- odbcConnect('tiny')
is_st(con, c('SH600000','SZ000001'), 20110104)
}

}
