% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/if_can_buy.R
\name{if_can_buy}
\alias{if_can_buy}
\title{Test if the stock can buy}
\usage{
if_can_buy(con, wind_code, buz_day, ...)
}
\arguments{
\item{con}{a tiny connection}

\item{wind_code}{vecter of future code list in wind_code format}

\item{buz_day}{int or char, format like \%Y\%m\%d, test for this day}

\item{full}{logi, return only wind_code can buy or total with tibble}

\item{delist_adj}{logi, True if delist stock canbuy}
}
\value{
a tibble with stock code and flag that 1 means purchasable and 0 not
}
\description{
Test if the stock can buy or not at business day
}
\details{
if the amount equal to 0 or high price equal to low price, the stock regards can't buy that day.
}
\examples{
\dontrun{
if_can_buy(con, c('600000.SH','000001.SH'), 20110104)
}

}
