% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_st.R
\name{is_st}
\alias{is_st}
\title{Test if the stock is st or not}
\usage{
is_st(con, wind_code, buz_day, ...)
}
\arguments{
\item{con}{a tiny connection}

\item{wind_code}{vecter of future code list in wind_code format}

\item{buz_day}{integer or character format like \%Y\%m\%d, test for this day}
}
\value{
a data.frame with stock code and flag that 1 means st and 0 not
}
\description{
Test if the stock is st or not at business day
}
\examples{
\dontrun{
is_st(con, c('600000.SH','000001.SH'), 20110104)
}

}
