% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_index_acount.R
\docType{data}
\name{m_index_acount}
\alias{m_index_acount}
\title{equal weight index}
\format{A data frame with 2638 rows and 2 variables}
\description{
equal weight to by the all stock which their first day > 90 and the tradeday > 5 and not st
the index change stock the first business day of each month
}
\details{
\itemize{
  \item date. date of index
  \item acount.
}
}
\keyword{datasets}
