% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/backtest.R
\docType{methods}
\name{marginals,backtest-method}
\alias{marginals,backtest-method}
\title{The \code{marginals} method returns a list of matrices, one matrix for each \code{in.var},
where the value of each cell is the number of observations for that \code{in.var} and
\code{by.var} combination. Different from \code{counts} because the marginal sums have been
appended to the matrices.}
\usage{
\S4method{marginals}{backtest}(object)
}
\description{
The \code{marginals} method returns a list of matrices, one matrix for each \code{in.var},
where the value of each cell is the number of observations for that \code{in.var} and
\code{by.var} combination. Different from \code{counts} because the marginal sums have been
appended to the matrices.
}

