% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/marginals.R
\docType{methods}
\name{marginals,backtest-method}
\alias{marginals,backtest-method}
\title{\code{marginals} Method}
\usage{
\S4method{marginals}{backtest}(object)
}
\arguments{
\item{object}{is the backtest object to work on}
}
\value{
\code{marginals} method returns a list of matrices, one matrix for each \code{in.var}
}
\description{
The \code{marginals} method returns the value of observations for \code{in.var} and
\code{by.var} combination. Different from \code{counts} because the marginal sums have been
appended to the matrices.
}
\examples{
data(starmine)
bt <- backtest(starmine, in.var = "smi", ret.var = "ret.0.1.m", by.period = FALSE)
marginals(bt)
}
\seealso{
\code{\link{backtest}} and \code{\link{backtest-class}} and \code{\link{summary}}
}

