% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_back.R
\name{max_back}
\alias{max_back}
\title{Max back}
\usage{
max_back(x, type = c("yield", "acount"))
}
\arguments{
\item{x}{yield vector}

\item{type}{yield for yield input, acount for net value input}
}
\description{
calculate the max back
}
\examples{
data(m_index_acount)
max_back(m_index_acount$acount)

}
