% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/backtest.R
\docType{methods}
\name{means,backtest-method}
\alias{means,backtest-method}
\title{The means method returns a list of matrices, with one matrix for each \code{in.var}, where the
value of each cell is the mean of the returns for that \code{in.var} and \code{by.var} combination.}
\usage{
\S4method{means}{backtest}(object)
}
\description{
The means method returns a list of matrices, with one matrix for each \code{in.var}, where the
value of each cell is the mean of the returns for that \code{in.var} and \code{by.var} combination.
}

