% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/backtest.R
\docType{methods}
\name{naCounts,backtest-method}
\alias{naCounts,backtest-method}
\title{The \code{naCounts} method returns a list of matrices, with one matrix for each \code{in.var},
where the value of each cell is the number of NA observations for that \code{in.var} and
\code{by.var} combination.}
\usage{
\S4method{naCounts}{backtest}(object)
}
\description{
The \code{naCounts} method returns a list of matrices, with one matrix for each \code{in.var},
where the value of each cell is the number of NA observations for that \code{in.var} and
\code{by.var} combination.
}

