% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/naCounts.R
\docType{methods}
\name{naCounts,backtest-method}
\alias{naCounts,backtest-method}
\title{\code{naCounts} Method}
\usage{
\S4method{naCounts}{backtest}(object)
}
\arguments{
\item{object}{is the backtest object to work on}
}
\value{
a list of matrices containing the number of NA observations
}
\description{
The \code{naCounts} method returns a list of matrices, with one matrix for each \code{in.var},
where the value of each cell is the number of NA observations for that \code{in.var} and
\code{by.var} combination.
}
\examples{
data(starmine)
bt <- backtest(starmine, in.var = "smi", ret.var = "ret.0.1.m", by.period = FALSE)
naCounts(bt)
}
\seealso{
\code{\link{backtest}} and \code{\link{backtest-class}} and \code{\link{summary}}
}

