% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_buz_day.R
\name{next_buz_day}
\alias{next_buz_day}
\title{find next business day}
\usage{
next_buz_day(con, x, including = F)
}
\arguments{
\item{con, }{rdf connection}

\item{x, }{integer, date list}

\item{including, }{logic, if include the today}
}
\description{
find the next business day
}
\examples{
x <- rep(20170101+ 1:2, 5)
next_buz_day(x)

}
