% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/backtest.R
\docType{methods}
\name{plot,backtest,missing-method}
\alias{plot,backtest,missing-method}
\title{The \code{plot} method plots returns, cumulative returns, or turnover, when passed a \code{type}
argument of \code{return}, \code{cumreturn}, or \code{turnover}, respectively.}
\usage{
\S4method{plot}{backtest,missing}(x, type = "return", ...)
}
\description{
The \code{plot} method plots returns, cumulative returns, or turnover, when passed a \code{type}
argument of \code{return}, \code{cumreturn}, or \code{turnover}, respectively.
}

