% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharpe.R
\name{sharpe}
\alias{sharpe}
\title{Sharpe ratio}
\usage{
sharpe(x, rf = 0, adjust = 1)
}
\arguments{
\item{x}{yield vector}

\item{rf}{free interest rate}

\item{adjust}{adjust for the stardard error date}
}
\description{
calculate the sharpe ratio
}
\examples{
data(m_index_acount)
sharpe(cal_yield(m_index_acount$acount))

}
