% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/backtest.R
\docType{methods}
\name{show,backtest-method}
\alias{show,backtest-method}
\title{The show method prints the variables used in this backtest.}
\usage{
\S4method{show}{backtest}(object)
}
\description{
The show method prints the variables used in this backtest.
}

