% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/show.R
\docType{methods}
\name{show,backtest-method}
\alias{show,backtest-method}
\title{\code{Show} method}
\usage{
\S4method{show}{backtest}(object)
}
\arguments{
\item{object}{is the backtest object we want summary of}
}
\description{
The \code{show} method prints a list of \code{.var} variables used in this backtest
}
\examples{
data(starmine)
bt <- backtest(starmine, in.var = "smi", ret.var = "ret.0.1.m", by.period = FALSE)
show(bt)
}
\seealso{
\code{\link{backtest}} and \code{\link{backtest-class}} and \code{\link{summary}}
}

