% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock_acount.R
\docType{class}
\name{stock_acount}
\alias{stock_acount}
\title{stock_acount}
\format{An \code{\link{R6Class}} generator object
\preformatted{
 stock_acount$new(con = NULL, acount_f = 1000000, fee = 0.005)

}}
\usage{
stock_acount
}
\arguments{
\item{con}{a tiny connection}

\item{acount_f}{integer, the initial money you have}

\item{fee}{float, the fee you sold stock}

\item{date}{date, the date you alter you stock acount}

\item{stock}{vector, the stock you buy or sell}

\item{num}{the stock vol you buy or sell}

\item{amount}{the stock amount you buy}
}
\description{
create a R6 object of stock_acount.
}
\details{
the stock acount only charge the fee when you sold stock.
one of num and amount is needed. if it have both, the num will be used
amount is prefered becasue the price adjusted can often be wrong.
}
\section{Details}{

\code{$new(con, fee = 0.005)} to create new object;
\code{set_connection}
\code{order_buy} to buy the stock with fix number or amount
\code{order_sell} to sell the stock with fix number
\code{acount_update} to update the net value
}

\examples{
\dontrun{
con <- odbcConnect('tiny')
my_acount <- stock_acount$new(con = con, acount_f = 1000000)
my_acount$order_buy(as.Date('2015-01-04'), stock = 'SH600000',num = 1000)
my_acount$acount_update(as.Date('2015-01-04'))
my_acount$order_sell(as.Date('2015-01-05'), stock = 'SH600000', num = 1000)
my_acount$order_to(date, stock, weight, amount = 'all')
my_acount$acount_update(as.Date('2015-01-05'))
my_acount$show_acount_f()
}

}
\keyword{data}
