% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/backtest.R
\docType{methods}
\name{summary,backtest-method}
\alias{summary,backtest-method}
\title{The summary method prints the results of the backtest.}
\usage{
\S4method{summary}{backtest}(object, ...)
}
\description{
The summary method prints the results of the backtest.
}

