% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summary.R
\docType{methods}
\name{summary,backtest-method}
\alias{summary,backtest-method}
\title{Summary method}
\usage{
\S4method{summary}{backtest}(object, ...)
}
\arguments{
\item{object}{is the backtest object we want summary of}

\item{...}{is the additional arguments to pass on}
}
\description{
The \code{summary} method prints the results of the backtest. It first prints a header listing
the \code{.vars} variables, then prints the table returned by the \code{spreads} method.
}
\note{
\code{summary} provides different displays depending on the type of \code{backtest} object.
}
\examples{
data(starmine)
bt <- backtest(starmine, in.var = "smi", ret.var = "ret.0.1.m", by.period = FALSE)
summary(bt)
}
\seealso{
\code{\link{backtest}} and \code{\link{backtest-class}}
}

