% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_acount.R
\name{summary_acount}
\alias{summary_acount}
\title{Summary the acount performance}
\usage{
summary_acount(total_acount, benchmark = F, ylog = FALSE, plot = T,
  show_result = T, ...)
}
\arguments{
\item{total_acount}{dataframe with daily net value. if benchmark mod there must be a variable named benchmark. All acount should stardardize for 1 million initial amount}

\item{benchmark}{logical, if relative mod or not}

\item{ylog}{logical, set the y-axis to logarithmic scale, default FALSE}

\item{plot}{logical, plot or not}

\item{...}{see details in \code{\link{charts.PerformanceSummary}}}
}
\value{
A plot show the cumulative return and drawdown
If there's no benchmark ,a dataframe includes that by year
\itemize{
 \item{\code{yeild}: last value this year / last value last year}
 \item{\code{max_back}: max back}
 \item{\code{back_ratio}: yeild / max_back}
 \item{\code{sharpratio}: mean(yeild)/(sd(yeild) * sqrt(the business day this year))}
 \item{\code{capital_loss}: min(0,min(net value)/1000000 - 1)}
}
If there's benchmark mode,a dataframe includes more by year that
\itemize{
 \item{\code{yeild_r}: cumprod(yeild - benchmark yeild)}
 \item{\code{max_back_r}: max back for yeild_r}
 \item{\code{back_ratio_r}: yeild_r / max_back_r}
 \item{\code{sharpratio}: mean(yeild_r)/(sd(yeild_r) * sqrt(the business day this year))}
}
}
\description{
Plot the acount performance and return perfomance indicator
}
\details{
when it has benchmark, input data must have a variable named benchmark, and it can only have two acount
when it benchmark is FALSE, input data can have a variable names benchmark but it will be ignore.
}
\examples{
data(m_index_acount)
summary_acount(m_index_acount, benchmark = FALSE)

}
