% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_sign.R
\name{summary_sign}
\alias{summary_sign}
\title{Summary account with sign}
\usage{
summary_sign(date, yield, sign, fee = 0.005, plot = T, relative = T)
}
\arguments{
\item{date}{the date of the acount}

\item{yield}{the yield each day}

\item{sign}{the sign of buy or not.}

\item{fee}{the fee when sell the stock}

\item{plot}{plot or not}

\item{relative}{if output relative yield or not}
}
\value{
a data.frame with yield, win_loss, win_rate, max back by year. and these indicate for whole period
}
\description{
summary the acount with sign
}
\details{
the sign can be regard as stock position. make sure that the sign can get at that day.
}
