% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/backtest.R
\docType{methods}
\name{summaryStats,backtest-method}
\alias{summaryStats,backtest-method}
\title{The summaryStats method returns a data frame with spreads for each \code{date.var} value and each
\code{in.var}.}
\usage{
\S4method{summaryStats}{backtest}(object, mean = FALSE)
}
\description{
The summaryStats method returns a data frame with spreads for each \code{date.var} value and each
\code{in.var}.
}

