% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summaryStats.R
\docType{methods}
\name{summaryStats,backtest-method}
\alias{summaryStats,backtest-method}
\title{\code{summaryStats} method}
\usage{
\S4method{summaryStats}{backtest}(object, mean = FALSE)
}
\arguments{
\item{object}{is the backtest object we want summary of}

\item{mean}{is a boolean value for whether to return a table of means by calling \code{.bt.mean}}
}
\value{
a data frame summarizing the results of the backtest.  The entries of the data frame
contain means in cases 1, 2, and 4, and spreads in cases 3 and 5.
}
\description{
The \code{summaryStats} method returns a data frame with spreads for each \code{date.var} value and each
\code{in.var}.
}
\note{
When a table of means is returned, the \code{.bt.spread} function is called and summary spread
      data is attached to the right side of the data frame.  If a date.var is used,
      \code{.bt.mean} is called and mean summary data are attached to bottom of the data frame.
}
\examples{
data(starmine)
bt <- backtest(starmine, in.var = "smi", ret.var = "ret.0.1.m", by.period = FALSE)
summaryStats(bt)
}
\seealso{
\code{\link{backtest}} and \code{\link{backtest-class}} and \code{\link{summary}}
}

