% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tday_offset.R
\name{tday_offset}
\alias{tday_offset}
\title{trade day offset}
\usage{
tday_offset(con, buz_day, offset = 0)
}
\arguments{
\item{con, }{rdf connection}

\item{buz_day, }{integer or date}

\item{offset, }{integer}
}
\description{
find the offset of trade day
}
\examples{
tday_offset(con, 20170101, -10)

}
