% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/backtest.R
\docType{methods}
\name{totalCounts,backtest-method}
\alias{totalCounts,backtest-method}
\title{The \code{totalCounts} method returns a data frame in the same format as the speads data frame
returned by \code{summaryStats}: contains the sum of counts for all buckets (or high and
low buckets if argument \code{low.high.only} is set to TRUE) of non-NA
\code{in.var} values that went into the spread calculations. It is different from counts because
it displays the sum of counts from all buckets (or lowest and highest only), thus
allowing for output that matches the format of spreads output.}
\usage{
\S4method{totalCounts}{backtest}(object, low.high.only = FALSE)
}
\description{
The \code{totalCounts} method returns a data frame in the same format as the speads data frame
returned by \code{summaryStats}: contains the sum of counts for all buckets (or high and
low buckets if argument \code{low.high.only} is set to TRUE) of non-NA
\code{in.var} values that went into the spread calculations. It is different from counts because
it displays the sum of counts from all buckets (or lowest and highest only), thus
allowing for output that matches the format of spreads output.
}

