% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/backtest.R
\docType{methods}
\name{turnover,backtest-method}
\alias{turnover,backtest-method}
\title{The \code{turnover} method returns a \code{data.frame} of the turnovers if the \code{backtest}
is \code{natural}. Passing a \code{mean} argument will append the mean of the turnover(s) as the
last row of the matrix}
\usage{
\S4method{turnover}{backtest}(object, mean = FALSE)
}
\description{
The \code{turnover} method returns a \code{data.frame} of the turnovers if the \code{backtest}
is \code{natural}. Passing a \code{mean} argument will append the mean of the turnover(s) as the
last row of the matrix
}

