% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/turnover.R
\docType{methods}
\name{turnover,backtest-method}
\alias{turnover,backtest-method}
\title{\code{turnover} Method}
\usage{
\S4method{turnover}{backtest}(object, mean = FALSE)
}
\arguments{
\item{object}{is the backtest object to work on}

\item{mean}{is a boolean value for whether append the mean of the turnover(s) as the last row of matrix}
}
\value{
a \code{data.frame} of the turnovers if the \code{backtest}
        is \code{natural}
}
\description{
The \code{turnover} method returns a \code{data.frame} of the turnovers if the \code{backtest}
is \code{natural}.
}
\examples{
data(starmine)
bt <- backtest(starmine, in.var = "smi", ret.var = "ret.0.1.m", by.period = FALSE, natural = TRUE,
               date.var = "date", id.var = "id")
turnover(bt)
}
\seealso{
\code{\link{backtest}} and \code{\link{backtest-class}} and \code{\link{summary}}
}

