% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/backtest.function.R
\name{backtest}
\alias{backtest}
\title{Returns Backtest Object}
\usage{
backtest(x, in.var, ret.var, universe, by.var = NULL, date.var = NULL,
  id.var = NULL, buckets = 5, natural = FALSE, do.spread = TRUE,
  by.period = TRUE, overlaps = 1)
}
\arguments{
\item{x}{is a data frame containing all raw data.}

\item{in.var}{is a character string or a vector of multiple character strings which correspond
to column names in the data frame \code{x}}

\item{ret.var}{is a character string or a vector of multiple character strings which correspond
to column names in the data frame \code{x}}

\item{universe}{is an expression for selecting a subset of the data frame \code{x}}

\item{by.var}{is a character string corresponding to column names in the data frame \code{x};
only one \code{by.var} is allowed}

\item{date.var}{is a character string corresponding to column names in the data frame \code{x};
only one \code{date.var} is allowed; date.var currently acts the same as a \code{by.var}}

\item{id.var}{is a character string corresponding to column names in the data frame \code{x};
it is only used in combination with \code{date.var} to uniquely identify stocks across dates}

\item{buckets}{is a numeric specifying the number of buckets in which to group \code{in.var}
and \code{by.var} respectively}

\item{natural}{}

\item{do.spread}{}

\item{by.period}{}

\item{overlaps}{}
}
\value{
a backtest object
}
\description{
\code{backtest.compute} returns a backtest object
}

