% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_acount.R
\docType{class}
\name{factor_acount}
\alias{factor_acount}
\title{factor_acount}
\format{\describe{
\item{\code{stock_acount$new(con, buy_stock)}}{to create new object}
\item{\code{factor_update(buy_num, begin_date, end_date, wait_sell = F, refresh = T, show_trace = T)}}{factor buy top buy_num stock every month}
\item{\code{show_wait_sell()}}{to show stock can't sell}
}}
\usage{
\preformatted{
p <- factor_acount$new(con, buy_stock = NULL)
p$factor_update(buy_num, begin_date, end_date, wait_sell = F, refresh = T, show_trace = T)
p$show_wait_sell()
}
}
\arguments{
\item{con}{connection for data}

\item{buy_stock}{data.frame, column code, buy_date(date), num(int)}

\item{begin_date}{int, the first date to backtest}

\item{end_date, }{int, the end date to backtest}

\item{wait_sell, }{logic. if True, sell the stock wait for sell every day}

\item{refresh, }{logic, clean acount history}

\item{show_trace, }{logic, print history day by day}

\item{...}{the paramater inherit from stock_acount}
}
\description{
create a R6 object of acount to analyse factor. can calculate the factor/number rank stock easily.
}
\details{
An \code{\link{R6Class}} generator object inherit from stock_acount
}
\examples{
\dontrun{
con <- odbcConnect('tiny')
my_acount <- factor_acount$new(con = con, 
                               buy_stock = data_all \%>\% 
                                  transmute(code = wind_code, buy_date = ymd(trade_dt)) \%>\%
                                  group_by(buy_date) \%>\% mutate(num = 1:n()))
my_acount$factor_update(20, 20170101, 20170531, wait_sell = T)
}

}
\keyword{data}
