% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/if_can_buy.R
\name{if_can_buy}
\alias{if_can_buy}
\alias{if_can_buy.default}
\alias{if_can_buy.tiny}
\alias{if_can_buy.rdf}
\title{Test if the stock can buy}
\usage{
if_can_buy(...)

\method{if_can_buy}{default}(...)

\method{if_can_buy}{tiny}(con, stocks, buz_day)

\method{if_can_buy}{rdf}(con, stocks, buz_day, full = F)
}
\arguments{
\item{con}{a tiny connection}

\item{stocks}{vecter of stock code list in tiny format}

\item{buz_day}{int or char, format like \%Y\%m\%d, test for this day}
}
\value{
a data.frame with stock code and flag that 1 means purchasable and 0 not
}
\description{
Test if the stock can buy or not at business day
}
\details{
if the amount equal to 0 or high price equal to low price, the stock regards can't buy that day.
}
\examples{
\dontrun{
con <- odbcConnect('tiny')
if_can_buy(con, c('SH600000','SZ000001'), 20110104)
}
}
