% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/overlaps.compute.R
\name{overlaps.compute}
\alias{overlaps.compute}
\title{Calculates the weighted returns}
\usage{
overlaps.compute(x, in.factor, date.var, id.var, overlaps)
}
\arguments{
\item{in.factor}{is the name of the vector of factors in \code{x} created
from the single \code{in.var} variable that the overlaps > 1 option allows}

\item{overlaps}{is the number of months a stock is held in a portfolio}
}
\value{
the numeric weighted returns for a portfolio
}
\description{
\code{overlaps.compute} Calculates the weighted returns for a long-short portfolio
corresponding to a backtest given the number of months a stock is held.
}
\details{
This measure of turnover ignores price changes
}

